Move dropdown stories to __docs__ folder
